# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "All-in-One Subsidy Platform API",
    "description": "API documentation for the All-in-One Subsidy Platform",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a new user account",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User registration information",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "username",
                "email",
                "password",
                "first_name",
                "last_name",
                "national_id",
                "date_of_birth",
                "address"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "first_name": {
                  "type": "string"
                },
                "last_name": {
                  "type": "string"
                },
                "national_id": {
                  "type": "string"
                },
                "date_of_birth": {
                  "type": "string",
                  "format": "date"
                },
                "address": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "user_id": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Authenticates a user and returns a token",
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "description": "User credentials",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed"
          }
        }
      }
    }
  }
}